\name{bucky-package}
\alias{bucky-package}
\alias{bucky}
\docType{package}
\title{
\packageTitle{bucky}
}
\description{
\packageDescription{bucky}
}
\details{
\packageDESCRIPTION{bucky}
\packageIndices{bucky}
}
\author{
\packageAuthor{bucky}

Maintainer: \packageMaintainer{bucky}
}
\references{
  Cameron, A. Colin, and Douglas L. Miller.
  \dQuote{\href{http://doi.org/10.3368/jhr.50.2.317}{A
      Practitioner's Guide to Cluster-Robust Inference}.}
  \emph{Journal of Human Resources} 50, no. 2 (Spring 2015): 317-372.
  
  Little, Roderick J. A., and Donald B. Rubin.
  \dQuote{\href{https://doi.org/10.1177/0049124189018002004}{The Analysis
      of Social Science Data with Missing Values}.}
  \emph{Sociological Methods & Research}
  18, no. 2-3 (November 1989): 292-326.

  Rubin, Donald. B. \emph{Multiple Imputation for Nonresponse in
    Surveys}. New York: John Wiley & Sons, 2004.
}
\keyword{ package }
\seealso{
  See also
  \code{\link[mice]{mice}},
  \code{\link[Amelia]{amelia-package}},
  \code{\link[mitools]{MIcombine}},
  \code{\link[sandwich]{sandwich}}
  and
  \code{\link[lmtest]{coeftest}}.
}
