\name{bucky-package}
\alias{bucky-package}
\alias{bucky}
\docType{package}
\title{
\packageTitle{bucky}
}
\description{
\packageDescription{bucky}
}
\details{
\packageDESCRIPTION{bucky}
\packageIndices{bucky}
}
\author{
\packageAuthor{bucky}

Maintainer: \packageMaintainer{bucky}
}
\references{
Cameron, A. Colin, and Douglas L. Miller. ``A Practitioner's Guide to
Cluster-Robust Inference.'' \url{http://cameron.econ.ucdavis.edu/research/Cameron_Miller_JHR_2015_February.pdf}
}
\keyword{ package }
\seealso{
See also \code{\link[sandwich]{sandwich}} and \code{\link[lmtest]{coeftest}}
}
